+++ conventions/forms/__init__.py
@@ -1,0 +1,10 @@
+from conventions.forms.avenant import *
+from conventions.forms.commentaires import *
+from conventions.forms.convention_number import *
+from conventions.forms.financement import *
+from conventions.forms.foyer_attribution import *
+from conventions.forms.foyer_variantes import *
+from conventions.forms.notification import *
+from conventions.forms.resiliation import *
+from conventions.forms.type1and2 import *
+from conventions.forms.upload import *
