+++ conventions/management/commands/clear_all_data.py
