+++ conventions/models/__init__.py
@@ -1,0 +1,6 @@
+from conventions.models.avenant_type import *
+from conventions.models.choices import *
+from conventions.models.convention import *
+from conventions.models.convention_history import *
+from conventions.models.piece_jointe import *
+from conventions.models.pret import *
