+++ conventions/models/convention.py
@@ -4,6 +4,9 @@ from django.db import models
 from django.db.models import Q
 from django.forms import model_to_dict
 from django.utils import timezone
+from conventions.models.avenant_type import AvenantType
+from conventions.models.choices import ConventionStatut, ConventionType1and2
+from conventions.models.convention_history import ConventionHistory
 
 from ecoloweb.models import EcoloReference
 from programmes.models import (
@@ -16,104 +19,6 @@ from programmes.models import (
     TypeStationnement,
 )
 from users.type_models import EmailPreferences, TypeRole
-
-
-class Preteur(models.TextChoices):
-    ETAT = "ETAT", "Etat"
-    EPCI = "EPCI", "EPCI"
-    REGION = "REGION", "Région"
-    VILLE = "VILLE", "Ville"
-    CDCF = "CDCF", "CDC pour le foncier"
-    CDCL = "CDCL", "CDC pour le logement"
-    COMMUNE = "COMMUNE", "Commune et action logement"
-    ANRU = "ANRU", "ANRU"
-    AUTRE = "AUTRE", "Autre"
-
-
-class ConventionStatut(models.TextChoices):
-    """
-    A/ PROJET : Projet - Création d'un projet de convention
-        Le bailleur crée un projet de convention APL, il ajoute à ce projet des documents
-        et des informations concernant les logements de l'opération.
-        Anciennement BROUILLON
-
-    B/ Instruction (2 sous-statuts) : Instruction du projet de convention
-        Le bailleur et l'instructeur échangent et s'alignent sur le contenu de la convention
-
-        B1/ INSTRUCTION : Instruction requise - Projet de convention soumis à l'instruction
-            L'instructeur vérifie le contenu du projet de convention et, si nécessaire,
-            transmet au bailleur ses demandes de modifications
-            Anciennement INSTRUCTION
-
-        B2/ CORRECTION : Corrections requises - Projet de convention à modifier par le bailleur
-            Le bailleur intègre les demandes de modification et soumet à nouveau le projet de
-            convention à l'instructeur
-            Anciennement CORRECTION
-
-    C/ A_SIGNER : A signer - Convention à signer
-        Le bailleur et l'instructeur sont d'accord sur le projet de convention. Les parties
-        procèdent à la signature.
-        Anciennement VALIDE
-
-    D/ SIGNEE : Signée - Convention signée
-        La convention signée est mise à disposition via la plateforme APiLos.
-        Anciennement CLOS
-
-    E/ RESILIEE: Résilié - Convention résiliée
-
-    F/ DENONCEE: Dénoncée - Convention dénoncée
-        Le bailleur décide, une fois la période de couverture de la convention achevée, de
-        dénoncer la convention et rendre ainsi caduques toutes les clauses d'encadrement qui
-        y étaient définies. Il est seul décideur, cependant depuis peu une commune considérée
-        comme en "carence" de logement social peut s'opposer à cette dénonciation.
-
-    G/ ANNULEE: Annulée en suivi - Convention annulée
-        Les 2 parties s'entendent pour annuler la convention
-
-    """
-
-    PROJET = "1. Projet", "Création d'un projet de convention"
-    INSTRUCTION = (
-        "2. Instruction requise",
-        "Projet de convention soumis à l'instruction",
-    )
-    CORRECTION = (
-        "3. Corrections requises",
-        "Projet de convention à modifier par le bailleur",
-    )
-    A_SIGNER = "4. A signer", "Convention à signer"
-    SIGNEE = "5. Signée", "Convention signée"
-    RESILIEE = "6. Résiliée", "Convention résiliée"
-    DENONCEE = "7. Dénoncée", "Convention dénoncée"
-    ANNULEE = "8. Annulée en suivi", "Convention annulée en suivi"
-
-    @classmethod
-    def active_statuts(cls):
-        return [cls.PROJET, cls.INSTRUCTION, cls.CORRECTION, cls.A_SIGNER]
-
-    @classmethod
-    def completed_statuts(cls):
-        return [cls.SIGNEE, cls.RESILIEE, cls.DENONCEE, cls.ANNULEE]
-
-
-class AvenantType(models.Model):
-    id = models.AutoField(primary_key=True)
-    nom = models.CharField(max_length=255, unique=True)
-    desc = models.TextField(null=True, blank=True)
-
-    def __str__(self):
-        return f"{self.nom}"
-
-    @classmethod
-    def get_as_choices(cls):
-        return [
-            (avenant_type.nom, avenant_type.nom) for avenant_type in cls.objects.all()
-        ]
-
-
-class ConventionType1and2(models.TextChoices):
-    TYPE1 = "Type1", "Type I"
-    TYPE2 = "Type2", "Type II"
 
 
 class Convention(models.Model):
@@ -677,151 +582,3 @@ class Convention(models.Model):
             if self.programme.administration
             else ""
         )
-
-
-class ConventionHistory(models.Model):
-    id = models.AutoField(primary_key=True)
-    uuid = models.UUIDField(default=uuid.uuid4, editable=False)
-    convention = models.ForeignKey(
-        "Convention",
-        on_delete=models.CASCADE,
-        null=False,
-        related_name="conventionhistories",
-    )
-    statut_convention = models.CharField(
-        max_length=25,
-        choices=ConventionStatut.choices,
-        default=ConventionStatut.PROJET,
-    )
-    statut_convention_precedent = models.CharField(
-        max_length=25,
-        choices=ConventionStatut.choices,
-        default=ConventionStatut.PROJET,
-    )
-    commentaire = models.TextField(null=True, blank=True)
-    user = models.ForeignKey(
-        "users.User",
-        related_name="valide_par",
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-    )
-    cree_le = models.DateTimeField(auto_now_add=True)
-    mis_a_jour_le = models.DateTimeField(auto_now=True)
-
-    # Needed for admin
-    @property
-    def bailleur(self):
-        return self.convention.programme.bailleur
-
-
-class Pret(models.Model):
-    id = models.AutoField(primary_key=True)
-    uuid = models.UUIDField(default=uuid.uuid4, editable=False)
-    convention = models.ForeignKey(
-        "Convention", on_delete=models.CASCADE, null=False, related_name="prets"
-    )
-    preteur = models.CharField(
-        max_length=25,
-        choices=Preteur.choices,
-        default=Preteur.AUTRE,
-    )
-    autre = models.CharField(null=True, blank=True, max_length=255)
-    date_octroi = models.DateField(null=True, blank=True)
-    numero = models.CharField(null=True, blank=True, max_length=255)
-    duree = models.IntegerField(null=True, blank=True)
-    montant = models.DecimalField(max_digits=12, decimal_places=2)
-    cree_le = models.DateTimeField(auto_now_add=True)
-    mis_a_jour_le = models.DateTimeField(auto_now=True)
-
-    # Needed to import xlsx files
-    import_mapping = {
-        "Numéro\n(caractères alphanuméric)": numero,
-        "Date d'octroi\n(format dd/mm/yyyy)": date_octroi,
-        "Durée\n(en années)": duree,
-        "Montant\n(en €)": montant,
-        "Prêteur\n(choisir dans la liste déroulante)": preteur,
-        "Préciser l'identité du préteur si vous avez sélectionné 'Autre'": autre,
-    }
-    sheet_name = "Financements"
-
-    # Needed for admin
-    @property
-    def bailleur(self):
-        return self.convention.programme.bailleur
-
-    def _get_preteur(self):
-        return self.get_preteur_display()
-
-    p = property(_get_preteur)
-
-    def _get_autre(self):
-        return self.autre
-
-    a = property(_get_autre)
-
-    def _get_date_octroi(self):
-        return self.date_octroi
-
-    do = property(_get_date_octroi)
-
-    def _get_numero(self):
-        return self.numero
-
-    n = property(_get_numero)
-
-    def _get_duree(self):
-        return self.duree
-
-    d = property(_get_duree)
-
-    def _get_montant(self):
-        return self.montant
-
-    m = property(_get_montant)
-
-    def p_full(self):
-        return self.get_preteur_display().replace(
-            "CDC", "Caisse de Dépôts et Consignation"
-        )
-
-    def preteur_display(self):
-        if self.preteur == Preteur.AUTRE:
-            return self.autre
-        return self.p_full()
-
-
-class PieceJointeType(models.TextChoices):
-    CONVENTION = "CONVENTION", "Convention APL"
-    RECTIFICATION = (
-        "RECTIFICATION",
-        "Demande de rectification(s) du bureau des hypothèques",
-    )
-    ATTESTATION_PREFECTORALE = "ATTESTATION_PREFECTORALE", "Attestations préfectorales"
-    AVENANT = "AVENANT", "Avenant"
-    PHOTO = "PHOTO", "Photographie du bâti ou des logements"
-    AUTRE = "AUTRE", "Autre"
-
-
-class PieceJointe(models.Model):
-    id = models.AutoField(primary_key=True)
-    uuid = models.UUIDField(default=uuid.uuid4, editable=False)
-    convention = models.ForeignKey(
-        "Convention",
-        on_delete=models.CASCADE,
-        null=False,
-        related_name="pieces_jointes",
-    )
-    type = models.CharField(
-        null=True,
-        max_length=24,
-        choices=PieceJointeType.choices,
-        default=PieceJointeType.AUTRE,
-    )
-    fichier = models.CharField(null=True, blank=True, max_length=255)
-    nom_reel = models.CharField(null=True, blank=True, max_length=255)
-    description = models.CharField(null=True, blank=True, max_length=255)
-    cree_le = models.DateTimeField(auto_now_add=False)
-
-    def is_convention(self) -> bool:
-        return self.type == PieceJointeType.CONVENTION
