+++ conventions/services/utils.py
@@ -121,21 +121,6 @@ def base_convention_response_error(request, convention):
     }
 
 
-def base_response_success(convention):
-    return {
-        "success": ReturnStatus.SUCCESS,
-        "convention": convention,
-    }
-
-
-def base_response_redirect_recap_success(convention):
-    return {
-        "success": ReturnStatus.SUCCESS,
-        "convention": convention,
-        "redirect": "recapitulatif",
-    }
-
-
 def editable_convention(request: HttpRequest, convention: Convention):
     if is_instructeur(request):
         return convention.statut in [
