+++ conventions/views/__init__.py
@@ -1,2 +1,16 @@
-from conventions.views.viewsconventions import *
-from conventions.views.viewsavenants import *
+from conventions.views.conventions import *
+from conventions.views.avenants import *
+from conventions.views.convention_form_annexes import *
+from conventions.views.convention_form_attribution import *
+from conventions.views.convention_form_bailleur import *
+from conventions.views.convention_form_cadastre import *
+from conventions.views.convention_form_collectif import *
+from conventions.views.convention_form_comments import *
+from conventions.views.convention_form_edd import *
+from conventions.views.convention_form_financement import *
+from conventions.views.convention_form_foyer_residence_logements import *
+from conventions.views.convention_form_logements import *
+from conventions.views.convention_form_programme import *
+from conventions.views.convention_form_selection import *
+from conventions.views.convention_form_type_stationnement import *
+from conventions.views.convention_form_variante import *
