+++ conventions/views/convention_form_comments.py
@@ -1,0 +1,11 @@
+from conventions.services.services_conventions import ConventionCommentsService
+from conventions.views.convention_form import ConventionView
+
+
+class ConventionCommentsView(ConventionView):
+    target_template: str = "conventions/comments.html"
+    service_class = ConventionCommentsService
+
+
+class AvenantCommentsView(ConventionCommentsView):
+    pass
