+++ conventions/views/convention_form_variante.py
@@ -1,0 +1,7 @@
+from conventions.services.variantes import ConventionFoyerVariantesService
+from conventions.views.convention_form import ConventionView
+
+
+class ConventionFoyerVariantesView(ConventionView):
+    target_template: str = "conventions/variantes.html"
+    service_class = ConventionFoyerVariantesService
